test: move provider action tests into own file
