+++ core/test/unit/src/actions.ts
@@ -42,151 +42,6 @@ describe("ActionRouter", () => {
     await garden.close()
   })
 
-  // Note: The test plugins below implicitly validate input params for each of the tests
-  describe("environment actions", () => {
-    describe("configureProvider", () => {
-      it("should configure the provider", async () => {
-        const config = { name: "test-plugin", foo: "bar", dependencies: [] }
-        const result = await actionRouter.provider.configureProvider({
-          ctx: await garden.getPluginContext(
-            providerFromConfig({
-              plugin: await garden.getPlugin("test-plugin"),
-              config,
-              dependencies: {},
-              moduleConfigs: [],
-              status: { ready: false, outputs: {} },
-            })
-          ),
-          namespace: "default",
-          environmentName: "default",
-          pluginName: "test-plugin",
-          log,
-          config,
-          configStore: garden.configStore,
-          projectName: garden.projectName,
-          projectRoot: garden.projectRoot,
-          dependencies: {},
-        })
-        expect(result).to.eql({
-          config,
-          moduleConfigs: [],
-        })
-      })
-    })
-
-    describe("augmentGraph", () => {
-      it("should return modules and/or dependency relations to add to the stack graph", async () => {
-        graph = await garden.getConfigGraph({ log: garden.log, emit: false })
-        const providers = await garden.resolveProviders(garden.log)
-        const result = await actionRouter.provider.augmentGraph({
-          log,
-          pluginName: "test-plugin",
-          actions: graph.getActions(),
-          providers,
-        })
-
-        const name = "added-by-test-plugin"
-        expect(result.addDependencies).to.eql([
-          {
-            by: {
-              kind: "Deploy",
-              name: "added-by-test-plugin",
-            },
-            on: {
-              kind: "Build",
-              name: "added-by-test-plugin",
-            },
-          },
-        ])
-        expect(result.addActions?.map((a) => ({ name: a.name, kind: a.kind }))).to.eql([
-          {
-            name: "added-by-test-plugin",
-            kind: "Build",
-          },
-          {
-            name: "added-by-test-plugin",
-            kind: "Deploy",
-          },
-        ])
-      })
-    })
-
-    describe("getDashboardPage", () => {
-      it("should resolve the URL for a dashboard page", async () => {
-        const page: DashboardPage = {
-          name: "foo",
-          title: "Foo",
-          description: "foodefoodefoo",
-          newWindow: false,
-        }
-        const result = await actionRouter.provider.getDashboardPage({ log, pluginName: "test-plugin", page })
-        expect(result).to.eql({
-          url: "http://foo",
-        })
-      })
-    })
-
-    describe("getEnvironmentStatus", () => {
-      it("should return the environment status for a provider", async () => {
-        const result = await actionRouter.provider.getEnvironmentStatus({ log, pluginName: "test-plugin" })
-        expect(result).to.eql({
-          ready: false,
-          outputs: {},
-        })
-      })
-    })
-
-    describe("prepareEnvironment", () => {
-      it("should prepare the environment for a configured provider", async () => {
-        const result = await actionRouter.provider.prepareEnvironment({
-          log,
-          pluginName: "test-plugin",
-          force: false,
-          status: { ready: true, outputs: {} },
-        })
-        expect(result).to.eql({
-          status: {
-            ready: true,
-            outputs: {},
-          },
-        })
-      })
-    })
-
-    describe("cleanupEnvironment", () => {
-      it("should clean up environment for a provider", async () => {
-        const result = await actionRouter.provider.cleanupEnvironment({ log, pluginName: "test-plugin" })
-        expect(result).to.eql({})
-      })
-    })
-
-    describe("getSecret", () => {
-      it("should retrieve a secret from the specified provider", async () => {
-        const result = await actionRouter.provider.getSecret({ log, pluginName: "test-plugin", key: "foo" })
-        expect(result).to.eql({ value: "foo" })
-      })
-    })
-
-    describe("setSecret", () => {
-      it("should set a secret via the specified provider", async () => {
-        const result = await actionRouter.provider.setSecret({
-          log,
-          pluginName: "test-plugin",
-          key: "foo",
-          value: "boo",
-        })
-        expect(result).to.eql({})
-      })
-    })
-
-    describe("deleteSecret", () => {
-      it("should delete a secret from the specified provider", async () => {
-        const result = await actionRouter.provider.deleteSecret({ log, pluginName: "test-plugin", key: "foo" })
-        expect(result).to.eql({ found: true })
-      })
-    })
-  })
-
   describe("getHandler", () => {
     const path = projectRootA
